Added response with status 404 on inexistent entries

+++ routes/contact_requests.js
@@ -2,15 +2,7 @@ const Joi = require('joi');
 const express = require('express');
 const router = express.Router();
 const db = require('../models');
-const {ValidationError} = require('sequelize');
-
-
-// const contact_request_schema = Joi.object({
-//     name: Joi.string(50).alphanum().min(3).max(50).required()
-// })
-
-
-
+const { ValidationError } = require('sequelize');
 
 
 router.get('', (req, res) => {
@@ -20,53 +12,81 @@ router.get('', (req, res) => {
 });
 
 router.get('/:id', (req, res) => {
-    if(db.Contact_requests)
     db.Contact_requests.findOne({
         where: {
-          id: req.params.id
+            id: req.params.id
         }
-      }).then(cr => res.send(cr));
+    }).then((cr) => {
+        if (cr) {
+            res.send(cr)
+        }
+        else {
+            res.status(404).send();
+        }
+    });
 });
 
-router.post('', async(req, res) => {
+router.post('', async (req, res) => {
     try {
         let item = await db.Contact_requests.build({
-            name: req.body.name, 
+            name: req.body.name,
             email: req.body.email,
             message: req.body.message,
             is_resolved: req.body.is_resolved
         });
         await item.validate();
-         // validation passed
-        await item.save(); 
+        // validation passed
+        await item.save();
         console.log('object created successfully');
         res.status(201).send();
-    }catch (e) {
-        if(e instanceof ValidationError){
+    } catch (e) {
+        if (e instanceof ValidationError) {
             res.status(400).send();
-            return console.error( 'Captured validation error: ', e.errors[0].message);
+            return console.error('Captured validation error: ', e.errors[0].message);
         }
         res.status(400).send();
     }
 });
 
 router.patch('/:id', (req, res) => {
-    db.Contact_requests.update({
-        is_resolved: req.body.is_resolved
-    },
-    {
+    db.Contact_requests.findOne({
         where: {
-            id:req.params.id
+            id: req.params.id
         }
-    }).then(() => res.status(200).send())
+    }).then((cr) => {
+        if (cr) {
+            db.Contact_requests.update({
+                is_resolved: req.body.is_resolved
+            },
+            {
+                where: {
+                    id: req.params.id
+                }
+            }).then(() => res.status(200).send());
+        }
+        else {
+            res.status(404).send();
+        }
+    });
 });
 
 router.delete('/:id', (req, res) => {
-    db.Contact_requests.destroy({
+    db.Contact_requests.findOne({
         where: {
-            id:req.params.id
+            id: req.params.id
         }
-    }).then(() => res.status(200).send());
+    }).then((cr) => {
+        if (cr) {
+            db.Contact_requests.destroy({
+                where: {
+                    id: req.params.id
+                }
+            }).then(() => res.status(200).send());
+        }
+        else {
+            res.status(404).send();
+        }
+    });
 });
 
 module.exports = router;
